fix: use int instead of quint64 for rate_limit to match logoscore -c parser

+++ logos-rln-module/src/i_logos_rln_module.h
@@ -28,14 +28,14 @@ public:
 
     /// Compute the rate commitment (leaf value) for a given id_commitment and rate_limit.
     /// Returns hex-encoded 32-byte rate commitment, or empty on failure.
-    virtual QString compute_rate_commitment(const QString& id_commitment_hex, quint64 rate_limit) = 0;
+    virtual QString compute_rate_commitment(const QString& id_commitment_hex, int rate_limit) = 0;
 
     /// Register a membership on the RLN tree.
     /// Returns JSON: {"leaf_index": N} on success, or empty on failure.
     virtual QString register_member(const QString& config_account_id,
                                     const QString& user_holding_account_id,
                                     const QString& id_commitment_hex,
-                                    quint64 rate_limit) = 0;
+                                    int rate_limit) = 0;
 };
 
 #define ILogosRlnModule_iid "org.logos.ilogosrlnmodule"
